+++ src/collective/z3cform/widgets/tests/test_acceptance.py
@@ -3,15 +3,13 @@
 from collective.z3cform.widgets.testing import FUNCTIONAL_TESTING
 from plone.testing import layered
 
+import os
 import robotsuite
 import unittest
 
-import pkg_resources
-PLONE_VERSION = pkg_resources.require("Plone")[0].version
-
-tests = [
-    'test_token_input_strip.txt',
-]
+dirname = os.path.dirname(__file__)
+files = os.listdir(dirname)
+tests = [f for f in files if f.startswith('test_') and f.endswith('.txt')]
 
 
 def test_suite():