+++ src/collective/z3cform/widgets/token_input_widget.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import zope.component
 import zope.interface
 import zope.schema
@@ -5,11 +7,43 @@ import zope.schema
 from z3c.form import interfaces
 from z3c.form import widget
 from z3c.form.browser import textarea
-
+from Products.Five.browser import BrowserView
 from zope.browserpage.viewpagetemplatefile import ViewPageTemplateFile
 
 from collective.z3cform.widgets.interfaces import ITokenInputWidget
 from collective.z3cform.widgets.interfaces import ILayer
+
+import json
+
+
+class ExportSubjectAsJSON(BrowserView):
+    """Return JSON search results for jQuery Tokeninput
+    for more information, see: http://loopj.com/jquery-tokeninput/#installation--setup
+    """
+
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def __call__(self):
+        self.request.response.setHeader("Content-type", "application/json")
+        if 'q' in self.request.keys():
+            query = self.request['q']
+            keys = self.context.portal_catalog.uniqueValuesFor('Subject')
+            keys = [k for k in keys if query in k]
+        else:
+            keys = []
+
+        # we will return up to 10 tokens only
+        tokens = map(self._tokenize, keys[:10])
+        return json.dumps(tokens)
+
+    def _tokenize(self, value):
+        if isinstance(value, str):
+            value = value.decode('utf-8')
+
+        return {'id': '%s' % value.replace(u"'", u"\\'"),
+                'name': '%s' % value.replace(u"'", u"\\'")}
 
 
 class TokenInputWidget(textarea.TextAreaWidget):
@@ -24,7 +58,7 @@ class TokenInputWidget(textarea.TextAreaWidget):
     js_template = u"""\
     (function($) {
         $().ready(function() {
-            var newValues = [%(newtags)s];
+            var newValues = '%(newtags)s';
             var oldValues = [%(oldtags)s];
             $('#%(id)s').data('klass','%(klass)s');
             keywordTokenInputActivate('%(id)s', newValues, oldValues);
@@ -35,19 +69,10 @@ class TokenInputWidget(textarea.TextAreaWidget):
     def js(self):
         if not ILayer.providedBy(self.request):
             return ""
-        values = self.context.portal_catalog.uniqueValuesFor('Subject')
         old_values = self.context.Subject()
-        tags = u""
         old_tags = u""
         index = 0
-        for index, value in enumerate(values):
-            if isinstance(value, str):
-                value = value.decode("utf-8")
-            tags += u"{id: '%s', name: '%s'}" % (
-                value.replace(u"'", u"\\'"), value.replace(u"'", u"\\'"))
-            if index < len(values) - 1:
-                tags += ", "
-
+        newtags = self.context.absolute_url() + "/json-subjects"
         #prepopulate
         for index, value in enumerate(old_values):
             if isinstance(value, str):
@@ -59,7 +84,7 @@ class TokenInputWidget(textarea.TextAreaWidget):
         result = self.js_template % dict(
             id=self.id,
             klass=self.klass,
-            newtags=tags,
+            newtags=newtags,
             oldtags=old_tags
         )
         return result